myapp/decorator_test.go
package myapp

import (
	"bufio"
	"bytes"
	"log"
	"net/http"
	"net/http/httptest"
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestIndexPage(t *testing.T) {
	assert := assert.New(t)

	ts := httptest.NewServer(NewDecoServer())
	defer ts.Close()

	res, err := http.Get(ts.URL)
	assert.NoError(err)
	assert.Equal(http.StatusOK, res.StatusCode)

	// data, _ := ioutil.ReadAll(res.Body)
}

func TestDecoHandler(t *testing.T) {
	assert := assert.New(t)

	ts := httptest.NewServer(NewDecoServer())
	defer ts.Close()

	buf := &bytes.Buffer{}
	log.SetOutput(buf) // default로 설정된 stdoutput 이 아닌 사용자가 지정한 buf 에 담기도록함.

	// index URL
	res, err := http.Get(ts.URL)
	assert.NoError(err)
	assert.Equal(http.StatusOK, res.StatusCode)

	r := bufio.NewReader(buf)
	line, _, err := r.ReadLine()
	assert.NoError(err)
	assert.Contains(string(line), "[LOGGER1] Started")
}
